Additional errors from avalara response are not included and handled
When the Avalara Webservice returns a status code error(500) the errors is not
well handled and just an exception is raised, and sometimes when the status code
is 200 but the street has some validation errors those errors are not injected
into the active records object for that address.
With the aim of managing those errors correctly some methods are added
in order to inject the errors into the right active record object and avoiding
the exception stopping de program flow execution.