DIG-1371 Tax does not match Spree
The logic to calculate the order discounts is not taking under
consideration all the adjustments of the order (e.g. Shipment discount),
also if the order includes a not eligible discount it gets throught
Avalara.

Another scenario is when the order includes an adjustment but just
before the customer place the order the promotion expires, the tax
adjustment is not updated.

- Refactor order discounts calculation to get all the adjustments of the
  order except the Tax and Line items adjustments (The line items
  adjustments are sent along with the line item)
- Refactor the calculation of the line item adjustment to only get the
  eligibles